refactor TradingEntityHit -> SearchHit
- rename to SearchHit and move into search/components
- use slots instead of conditional blocks for advanced layout
- use CSS vars for style overrides (reduce dependence on :global())
- extract SearchHitDescription
