+++ src/lib/search/index.ts
@@ -1,0 +1,5 @@
+export { default as searchClient } from './client';
+export { default as tradingEntities } from './trading-entities';
+export { default as Search } from './components/Search.svelte';
+export { default as SearchHit } from './components/SearchHit.svelte';
+export { default as SearchHitDescription } from './components/SearchHitDescription.svelte';
